Add access to meta data about client error (#25)
We have a possibility to set meta information about the error.
However, for some reasons we don't have ability to read it directly.
This commit adds getter for `code` attribute of `ClientError`.

This commit also updates the context provided to writeResponse()
method. Now it respects any updates of the context made by
middleware functions.